+++ easymode/i18n/__init__.py
@@ -19,7 +19,7 @@ from weakref import WeakKeyDictionary
 
 from django.db.models.signals import post_save
 
-from easymode.i18n.gettext import MakeModelMessages
+import easymode.i18n.gettext
 
 __all__ = ('register', 'unregister', 'admin', 'decorators', 'gettext', 'meta')
 
@@ -54,8 +54,8 @@ def register(cls, location=None):
         or a file name. If left unspecified, ``settings.LOCALE_DIR`` or\
         ``settings.PROJECT_DIR`` will be used.
     """
-       
-    create_po_for_model = MakeModelMessages(location, cls)
+    
+    create_po_for_model = easymode.i18n.gettext.MakeModelMessages(location, cls)
     
     _post_save_handlers[cls] = create_po_for_model