+++ easymode/i18n/admin/generic.py
@@ -1,6 +1,9 @@
-from django.forms.models import save_instance
 from django.contrib.contenttypes.generic import BaseGenericInlineFormSet
 from django.contrib.contenttypes.models import ContentType
+from django.forms.models import save_instance
+
+
+__all__ = ('LocalizableGenericInlineFormSet',)
 
 
 class LocalizableGenericInlineFormSet(BaseGenericInlineFormSet):