+++ easymode/tests/__init__.py
@@ -10,14 +10,13 @@ import sys
 
 from django.conf import settings
 
+from easymode.easypublisher.utils import filter_unpublished
 from easymode.utils import first_match, bases_walker, url_add_params
 from easymode.utils.languagecode import get_language_codes,\
     get_language_codes_as_disjunction, get_language_code_from_shorthand,\
     localize_fieldnames, get_real_fieldname, strip_language_code,\
     get_short_language_codes
 from easymode.utils.standin import standin_for
-from easymode.easypublisher.utils import filter_unpublished
-
 
 # check if some required settings are fulfilled
 if 'de' not in get_language_codes():