+++ docs/changes.rst
@@ -1,6 +1,50 @@
 Release Notes
 =============
 
+v0.14.0
+-------
+
+Added nofollow option to mark foreign keys that shouldn't be followed by the
+serializer. Nofollow can be used to optimize easymodes queries when generating
+xml, see :ref:`nofollow`. 
+
+v0.13.7
+-------
+
+- Fallbacks for translatable fields now also work when the first fallback is not
+  the MSGID_LANGUAGE.
+
+v0.13.6
+-------
+
+- You can now override the model form of an admin class decorated with L10n, just
+  like normal admin classes.
+
+v0.13.5
+-------
+
+- Easymode nolonger patches SubFieldBase. Fields that throw Exceptions when their
+  descriptor is accessed can now also be internationalized using I18n. This
+  includes ImageField and FileField.
+
+v0.13.4
+-------
+
+- :func:`~easymode.utils.standin.standin_for` now returns a standin that can be
+  pickled and unpickled.
+
+v0.13.3
+-------
+
+- register_all will nolonger try to register abstract models
+- search_fields is now supported for ModelAdmin classes that use L10n, however it
+  will not let you access related items.
+- You can now use fieldsets with the *Can edit untranslated fields* permission.
+- Added support for creating new objects to easypublisher.
+- Added tools to build preview functionality for drafts.
+- Added filter that removes unpublished items from the xml.
+- fixed error 'cannot import name introspection' caused by a circular import.
+
 v0.10.5
 -------